new file:   excel/test.xls 	new file:   excel/xlrdTest.py 	new file:   excel/xlwtTest.py 	new file:   html/MyHTMLParser.py 	new file:   html/tmpHtml.py 	modified:   urlsession/HttpClient.py 	modified:   urlsession/HttpClient.pyc 	new file:   urlsession/MyHTMLParser.py 	new file:   urlsession/MyHTMLParser.pyc 	modified:   urlsession/cl.py 	new file:   urlsession/clThread.py 	new file:   urlsession/clThread2.py 	modified:   urlsession/url.py

+++ urlsession/HttpClient.py
@@ -1,3 +1,4 @@
+#coding=utf-8
 import cookielib, urllib, urllib2
 import os
 
@@ -51,12 +52,15 @@ class HttpClient:
       print localPath
       os.mkdir(localPath)
     try:
-      
       filename = os.path.basename(sourceFile)
       localPathFileName = localPath+"/"+filename
-      print filename
-      return urllib.urlretrieve(sourceFile,localPathFileName)
+      if  os.path.exists(localPathFileName):
+        print ".....exists...."
+      else:
+        print filename
+        return urllib.urlretrieve(sourceFile,localPathFileName)
     except Exception as e:
+      print u"文件或者目录出错，请检查"
       return e
 
 #self.__cookie.clear() clean cookie
